Packages update

+++ GryphonUtilityBot.Web/Models/BotSingleton.cs
@@ -20,7 +20,7 @@ namespace GryphonUtilityBot.Web.Models
             }
             if ((config.AdminIds == null) || (config.AdminIds.Count == 0))
             {
-                config.AdminIds = JsonConvert.DeserializeObject<List<int>>(config.AdminIdsJson);
+                config.AdminIds = JsonConvert.DeserializeObject<List<long>>(config.AdminIdsJson);
             }
             Bot = new Bot.Bot(config);
         }
